NO-JIRA: Fix web terminal e2e race condition on operator lifecycle
Move operator install to a project-level setup (webterminal.setup.ts)
so parallel workers no longer race on beforeAll/afterAll hooks, causing
OAuth session loss and 409 conflicts on subscription creation.
