+++ frontend/e2e/tests/webterminal/developer/web-terminal-devuser.spec.ts
@@ -3,10 +3,6 @@ import type { Page } from '@playwright/test';
 import { test, expect } from '../../../fixtures';
 import type KubernetesClient from '../../../clients/kubernetes-client';
 import { WebTerminalPage } from '../../../pages/web-terminal-page';
-import {
-  ensureWebTerminalOperatorInstalled,
-  uninstallWebTerminalOperator,
-} from '../utils/web-terminal-operator';
 
 const DEVWORKSPACE_GROUP = 'workspace.devfile.io';
 const DEVWORKSPACE_VERSION = 'v1alpha2';
@@ -38,17 +34,9 @@ async function verifyDevWorkspaceRunning(
 }
 
 test.describe('Web Terminal for Developer user', () => {
-  test.beforeAll(async ({ k8sClient }) => {
-    await ensureWebTerminalOperatorInstalled(k8sClient);
-  });
-
   test.beforeEach(async ({ k8sClient, cleanup }) => {
     await k8sClient.createNamespace(EXISTING_PROJECT);
     cleanup.trackNamespace(EXISTING_PROJECT);
-  });
-
-  test.afterAll(async ({ k8sClient }) => {
-    await uninstallWebTerminalOperator(k8sClient);
   });
 
   test(
